requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
cryptography==41.0.7
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
Pillow==10.2.0